# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from "./userUpdateRole";

export interface UserUpdate {
  full_name?: string;
  username?: string;
  role?: UserUpdateRole;
  password?: string;
}